requirements.txt
azure-ai-agentserver-agentframework==1.0.0b14
azure-ai-agentserver-core==1.0.0b14
opentelemetry-semantic-conventions-ai==0.4.13
debugpy
# Agent development CLI tool (preview)
agent-dev-cli